Update the memory type choose logic and polish some code.
1. Change DataFormat from enum to enum class.

+++ mace/core/runtime/opencl/opencl_util.cc
@@ -173,7 +173,7 @@ void OpenCLUtil::BuildTransformOpDef(
   arg->set_i(static_cast<int32_t>(dt));
   arg = op_def->add_arg();
   arg->set_name("data_format");
-  arg->set_i(data_format);
+  arg->set_i(static_cast<int>(data_format));
   if (!input_shape.empty()) {
     OutputShape *shape = op_def->add_output_shape();
     for (auto value : input_shape) {
